backend/requirements.txt
# Face Recognition Backend Requirements
# Compatible with Python 3.12

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Face Recognition & Computer Vision
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy>=1.26.0  # Python 3.12 compatible
Pillow==10.2.0

# Liveness Detection (using scikit-learn)
scikit-learn==1.4.0
scipy==1.12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Validation
pydantic==2.5.3

# ASGI server
click==8.1.7
h11>=0.8

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3  # Pin to 4.1.3 (5.x has breaking changes)

# Export functionality
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.9
